extract auth info suite
